Initial build for PREPARE-TIMES-NZ (#78)
This is the first version of a method for the proposed data_raw -> data_intermediate workflow in PREPARE-TIMES-NZ, and the user config files that define how data should sit in the generated excel files to be read by Veda. It contains documentation on the methods and structures, and the scripts in PREPARE-TIMES-NZ/scripts/stage_2_baseyear act as an initial template for how these might work for other components of TIMES-NZ in future. 

We note that there is a bit of tidying to do on this first pass.